Add sweetalert

+++ src/app/book-list/book-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
+import Swal from 'sweetalert2';
 import { Book } from '../models/book.model';
 import { BooksService } from '../services/books.service';
 
@@ -30,8 +31,32 @@ export class BookListComponent implements OnInit {
   }
 
   onDeleteBook(book: Book) {
-    this.booksService.removeBook(book);
+    Swal.fire({
+
+      title: 'Are you sure want to remove?',
+      text: 'You will not be able to recover this data!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+        this.booksService.removeBook(book);
+        Swal.fire(
+          'Deleted!',
+          'Book data has been deleted',
+          'success'
+        )
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Book data is safe',
+          'error'
+        )
+      }
+    })
   }
+  
 
   onViewBook(id: number) {
     this.router.navigate(['/books', 'view', id]);
